+++ extensions/broker-server-api/api/src/main/java/de/sovity/edc/ext/brokerserver/api/BrokerServerResource.java
@@ -20,7 +20,7 @@ import de.sovity.edc.ext.brokerserver.api.model.ConnectorDetailPageQuery;
 import de.sovity.edc.ext.brokerserver.api.model.ConnectorDetailPageResult;
 import de.sovity.edc.ext.brokerserver.api.model.ConnectorPageQuery;
 import de.sovity.edc.ext.brokerserver.api.model.ConnectorPageResult;
-import de.sovity.edc.ext.brokerserver.api.model.DataOfferCountResult;
+import de.sovity.edc.ext.brokerserver.api.model.AuthorityPortalConnectorInfo;
 import de.sovity.edc.ext.brokerserver.api.model.DataOfferDetailPageQuery;
 import de.sovity.edc.ext.brokerserver.api.model.DataOfferDetailPageResult;
 import io.swagger.v3.oas.annotations.Operation;
@@ -81,9 +81,9 @@ public interface BrokerServerResource {
     void deleteConnectors(List<String> endpoints, @QueryParam("adminApiKey") String adminApiKey);
 
     @POST
-    @Path("authority-portal-api/data-offer-counts")
+    @Path("authority-portal-api/connectors")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    @Operation(description = "Query the amount of public Data Offers by provided Connector URLs")
-    DataOfferCountResult dataOfferCount(List<String> endpoints);
+    @Operation(description = "Provide Connector metadata by provided Connector Endpoints")
+    List<AuthorityPortalConnectorInfo> getConnectorMetadata(List<String> endpoints, @QueryParam("adminApiKey") String adminApiKey);
 }